Mixin TS2
