+++ .storybook/main.ts
@@ -57,6 +57,8 @@ const config: StorybookConfig = {
       },
     },
   },
+
+  docs: {},
 }
 
 export default config
