+++ .storybook/preview.tsx
@@ -191,7 +191,7 @@ const preview: Preview = {
     },
   },
 
-  tags: ['autodocs'],
+  tags: ['autodocs', 'autodocs'],
 }
 
 export default preview
