+++ src/components/Table/TableComponents/index.ts
@@ -1,0 +1,4 @@
+export { TableHeader } from './TableHeader'
+export { TableLoading } from './TableLoading'
+export { TableRows } from './TableRows'
+export { TableToolbar } from './TableToolbar'
