+++ src/components/Table/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useTableSearch } from './useTableSearch'
+export { useTableSort } from './useTableSort'
