+++ src/components/Table/utils/formatters.tsx
@@ -25,6 +25,5 @@ export const formatCellValue = (value: unknown): React.ReactNode => {
       </Box>
     )
   }
-
   return String(value)
 }
